requirements.txt
streamlit
pypdf
groq
sentence-transformers
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
environs==9.5.0
grpcio==1.63.0
idna==3.11
marshmallow==3.26.2
milvus-lite==2.4.12
numpy==2.4.3
orjson==3.11.7
packaging==26.0
pandas==3.0.1
protobuf==7.34.0
pymilvus==2.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
setuptools==69.5.1
six==1.17.0
tqdm==4.67.3
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
